+++ dart/geobase/lib/geometric.dart
@@ -1,0 +1,9 @@
+/// Geometric functions for cartesian calculations.
+///
+/// This libary exports a subset of `package:geobase/geobase.dart`.
+///
+/// Usage: import `package:geobase/geometric.dart`
+library geometric;
+
+export 'src/geometric/base/distanced_position.dart';
+export 'src/geometric/cartesian/areal/cartesian_areal_extension.dart';